Remove util module
The `util` module now contains only two private macros, move them to
`internal_macros` and remove the `util` module.
