+++ src/internal_macros.rs
@@ -2,6 +2,50 @@
 
 //! Non-public macros
 
+/// Adds `AsRef` implementation to a given type `$ty`.
+macro_rules! as_ref_impl(
+    ($ty:ident) => (
+        $crate::internal_macros::as_ref_impl!($ty, );
+    );
+    ($ty:ident, $($gen:ident: $gent:ident),*) => (
+        impl<$($gen: $gent),*> $crate::_export::_core::convert::AsRef<[u8]> for $ty<$($gen),*>  {
+            fn as_ref(&self) -> &[u8] { &self[..] }
+        }
+    )
+);
+pub(crate) use as_ref_impl;
+
+/// Adds an implementation of the `HashEngine::input` method.
+macro_rules! engine_input_impl(
+    ($n:literal) => (
+        #[cfg(not(hashes_fuzz))]
+        fn input(&mut self, mut inp: &[u8]) {
+            while !inp.is_empty() {
+                let buf_idx = self.length % <Self as crate::HashEngine>::BLOCK_SIZE;
+                let rem_len = <Self as crate::HashEngine>::BLOCK_SIZE - buf_idx;
+                let write_len = $crate::_export::_core::cmp::min(rem_len, inp.len());
+
+                self.buffer[buf_idx..buf_idx + write_len]
+                    .copy_from_slice(&inp[..write_len]);
+                self.length += write_len;
+                if self.length % <Self as crate::HashEngine>::BLOCK_SIZE == 0 {
+                    self.process_block();
+                }
+                inp = &inp[write_len..];
+            }
+        }
+
+        #[cfg(hashes_fuzz)]
+        fn input(&mut self, inp: &[u8]) {
+            for c in inp {
+                self.buffer[0] ^= *c;
+            }
+            self.length += inp.len();
+        }
+    )
+);
+pub(crate) use engine_input_impl;
+
 macro_rules! arr_newtype_fmt_impl {
     ($ty:ident, $bytes:expr $(, $gen:ident: $gent:ident)*) => {
         impl<$($gen: $gent),*> $crate::_export::_core::fmt::LowerHex for $ty<$($gen),*> {
@@ -68,7 +112,7 @@ macro_rules! hash_trait_impls {
 
         $crate::internal_macros::arr_newtype_fmt_impl!(Hash, $bits / 8 $(, $gen: $gent)*);
         serde_impl!(Hash, $bits / 8 $(, $gen: $gent)*);
-        as_ref_impl!(Hash $(, $gen: $gent)*);
+        $crate::internal_macros::as_ref_impl!(Hash $(, $gen: $gent)*);
 
         impl<$($gen: $gent),*> $crate::_export::_core::convert::AsRef<[u8; $bits / 8]> for Hash<$($gen),*> {
             fn as_ref(&self) -> &[u8; $bits / 8] { &self.0 }
