+++ src/ripemd160.rs
@@ -39,7 +39,7 @@ impl crate::HashEngine for HashEngine {
     #[inline]
     fn n_bytes_hashed(&self) -> usize { self.length }
 
-    engine_input_impl!(20);
+    crate::internal_macros::engine_input_impl!(20);
 
     #[cfg(not(hashes_fuzz))]
     #[inline]
